requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2
django-celery-results==2.5.1
django-ckeditor==6.5.1
django-crontab==0.7.1
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-mptt-admin==2.5.0
django-nested-admin==4.0.2
django-redis==5.3.0
django-taggit==4.0.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
env==0.1.0
environs==9.5.0
idna==3.4
inflection==0.5.1
kombu==5.3.0
marshmallow==3.19.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-monkey-business==1.0.0
python3-openid==3.2.0
pytils==0.4.1
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
rules==3.3
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
validate-email==1.3
vine==5.0.0
wcwidth==0.2.6
